Renamed the project from "Schedulr" to "Listr" and added in the templates/sample files.
