+++ src/main/java/com/natebeckemeyer/projects/listrgui/implementations/Completed.java
@@ -1,10 +1,10 @@
-package com.natebeckemeyer.projects.schedulrgui.implementations;
+package com.natebeckemeyer.projects.listrgui.implementations;
 
-import com.natebeckemeyer.projects.schedulrgui.core.AbstractTask;
-import com.natebeckemeyer.projects.schedulrgui.core.Rule;
+import com.natebeckemeyer.projects.listrgui.core.AbstractTask;
+import com.natebeckemeyer.projects.listrgui.core.Rule;
 
 /**
- * Created for Schedulr by @author Nate Beckemeyer on 2016-06-20.
+ * Created for Listr by @author Nate Beckemeyer on 2016-06-20.
  * <p>
  * This {@link Rule} returns true for all predicates for which {@link AbstractTask#isCompleted()} returns true.
  */
