+++ src/main/java/com/natebeckemeyer/projects/listrgui/implementations/Week.java
@@ -1,13 +1,13 @@
-package com.natebeckemeyer.projects.schedulrgui.implementations;
+package com.natebeckemeyer.projects.listrgui.implementations;
 
-import com.natebeckemeyer.projects.schedulrgui.core.AbstractTask;
-import com.natebeckemeyer.projects.schedulrgui.core.Rule;
+import com.natebeckemeyer.projects.listrgui.core.AbstractTask;
+import com.natebeckemeyer.projects.listrgui.core.Rule;
 
 import java.util.Calendar;
 import java.util.GregorianCalendar;
 
 /**
- * Created for Schedulr by @author Nate Beckemeyer on 2016-06-08.
+ * Created for Listr by @author Nate Beckemeyer on 2016-06-08.
  */
 public class Week implements Rule
 {
