+++ src/main/resources/samples/defaultProperties.properties
@@ -1,0 +1,5 @@
+#Wed Jul 13 20:35:22 CDT 2016
+auto-load=true
+auto-save-file=user/tasks/current.dat
+auto-save=true
+completion-behavior=SimpleCompleted
