+++ src/main/resources/system/templates/Rule.template
@@ -1,0 +1,15 @@
+$$$IMPORTS$$$
+
+public class $$$CLASSNAME$$$ $$$EXTENDS$$$ implements Rule$$$IMPLEMENTATIONS$$$
+{
+
+    @Override public String toString()
+    {
+        return "$$$NAME$$$";
+    }
+
+    @Override public boolean test(AbstractTask task)
+    {
+        $$$CODE$$$
+    }
+}
