+++ src/main/resources/system/windows/addBehaviorPopup.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Font?>
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="639.0"
             prefWidth="522.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="com.natebeckemeyer.projects.schedulrgui.graphics.AddBehaviorPopupController">
+            fx:controller="com.natebeckemeyer.projects.listrgui.graphics.AddBehaviorPopupController">
     <children>
         <TabPane layoutX="2.0" layoutY="-2.0" prefHeight="639.0" prefWidth="522.0" tabClosingPolicy="UNAVAILABLE">
             <tabs>
