+++ src/main/resources/system/windows/addTaskPopup.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="223.0"
             prefWidth="391.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="com.natebeckemeyer.projects.schedulrgui.graphics.AddTaskPopupController">
+            fx:controller="com.natebeckemeyer.projects.listrgui.graphics.AddTaskPopupController">
     <children>
         <Button layoutX="210.0" layoutY="188.0" mnemonicParsing="false"
                 onMouseClicked="#addTaskButtonClicked" text="Add Task"/>
