+++ src/main/resources/system/windows/mainWindow.fxml
@@ -2,7 +2,7 @@
 
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.AnchorPane?>
-<AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="613.0" prefWidth="999.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.natebeckemeyer.projects.schedulrgui.graphics.MainWindowController">
+<AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="613.0" prefWidth="999.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.natebeckemeyer.projects.listrgui.graphics.MainWindowController">
     <children>
         <TableView fx:id="mainTaskList" editable="true" layoutX="128.0" layoutY="86.0" onKeyPressed="#deleteTask" prefHeight="527.0" prefWidth="871.0">
             <columns>
